pyproject.toml
[tool.poetry]
name = "dungeon-design"
version = "0.1.0"
description = ""
authors = ["ratilmii <ratilmii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame-ce = "^2.5.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
